Update memory set

+++ src/main.rs
@@ -33,8 +33,7 @@ fn kernel_init() {
         fn ebss();
     }
     unsafe {
-        core::slice::from_raw_parts_mut(sbss as usize as *mut u8, ebss as usize - sbss as usize)
-            .fill(0);
+        core::slice::from_raw_parts_mut(sbss as *mut u8, ebss as usize - sbss as usize).fill(0);
     }
 
     misc::logger::init();
